Update int to int16_t for XYDrawComposited and XYDrawGradient

+++ src/fl/draw_visitor.h
@@ -14,7 +14,7 @@ namespace fl {
 // Draws a uint8_t value to a CRGB array, blending it with the existing color.
 struct XYDrawComposited {
     XYDrawComposited(const CRGB &color, const XYMap &xymap, CRGB *out);
-    void draw(const vec2<int> &pt, uint32_t index, uint8_t value);
+    void draw(const vec2<int16_t> &pt, uint32_t index, uint8_t value);
     const CRGB mColor;
     const XYMap mXYMap;
     CRGB *mOut;
@@ -22,7 +22,7 @@ struct XYDrawComposited {
 
 struct XYDrawGradient {
     XYDrawGradient(const Gradient &gradient, const XYMap &xymap, CRGB *out);
-    void draw(const vec2<int> &pt, uint32_t index, uint8_t value);
+    void draw(const vec2<int16_t> &pt, uint32_t index, uint8_t value);
     const Gradient mGradient;
     const XYMap mXYMap;
     CRGB *mOut;
@@ -32,7 +32,7 @@ inline XYDrawComposited::XYDrawComposited(const CRGB &color, const XYMap &xymap,
                                           CRGB *out)
     : mColor(color), mXYMap(xymap), mOut(out) {}
 
-inline void XYDrawComposited::draw(const vec2<int> &pt, uint32_t index,
+inline void XYDrawComposited::draw(const vec2<int16_t> &pt, uint32_t index,
                                    uint8_t value) {
     FASTLED_UNUSED(pt);
     CRGB &c = mOut[index];
@@ -45,7 +45,7 @@ inline XYDrawGradient::XYDrawGradient(const Gradient &gradient,
                                       const XYMap &xymap, CRGB *out)
     : mGradient(gradient), mXYMap(xymap), mOut(out) {}
 
-inline void XYDrawGradient::draw(const vec2<int> &pt, uint32_t index,
+inline void XYDrawGradient::draw(const vec2<int16_t> &pt, uint32_t index,
                                  uint8_t value) {
     FASTLED_UNUSED(pt);
     CRGB c = mGradient.colorAt(value);
